+++ Tests/PeripheryTests/Syntax/FunctionVisitTest.swift
@@ -69,7 +69,8 @@ class FunctionVisitTest: XCTestCase {
     func testFunctionWithGenericArguments() throws {
         let result = results[fixtureLocation(line: 20)]
         XCTAssertEqual(result?.parameterTypeLocations, [
-            fixtureLocation(line: 20, column: 70)
+            fixtureLocation(line: 20, column: 70),
+            fixtureLocation(line: 20, column: 72)
         ])
         XCTAssertEqual(result?.genericParameterLocations, [
             fixtureLocation(line: 20, column: 37),
@@ -105,7 +106,8 @@ class FunctionVisitTest: XCTestCase {
     func testInitializerWithGenericArguments() throws {
         let result = results[fixtureLocation(line: 30, column: 5)]
         XCTAssertEqual(result?.parameterTypeLocations, [
-            fixtureLocation(line: 30, column: 46)
+            fixtureLocation(line: 30, column: 46),
+            fixtureLocation(line: 30, column: 48)
         ])
         XCTAssertEqual(result?.genericParameterLocations, [
             fixtureLocation(line: 30, column: 13),
@@ -133,7 +135,8 @@ class FunctionVisitTest: XCTestCase {
     func testSubscriptWithGenericArguments() throws {
         let result = results[fixtureLocation(line: 35, column: 5)]
         XCTAssertEqual(result?.parameterTypeLocations, [
-            fixtureLocation(line: 35, column: 51)
+            fixtureLocation(line: 35, column: 51),
+            fixtureLocation(line: 35, column: 53)
         ])
         XCTAssertEqual(result?.returnTypeLocations, [
             fixtureLocation(line: 35, column: 62)