+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Workaround Swift shorthand if-let syntax bug (https://github.com/apple/swift/issues/61509). Global properties are not handled by this workaround.
 - Fix retaining inferred associated types.
 - Fix redundant public accessibility analysis for types used in closure signatures.
+- Conflicting index store units are now detected and result in an error.
 
 ## 2.10.0 (2022-10-10)