+++ CHANGELOG.md
@@ -10,6 +10,20 @@
 
 ##### Bug Fixes
 
+- None.
+
+## 2.10.1 (2022-11-20)
+
+##### Breaking
+
+- None.
+
+##### Enhancements
+
+- None.
+
+##### Bug Fixes
+
 - Fix error building with SwiftPM and Swift 5.7: `the target 'PeripheryKit' in product 'periphery' contains unsafe build flags`
 - Fix unused parameter analysis for shorthand if-let syntax.
 - Workaround Swift shorthand if-let syntax bug (https://github.com/apple/swift/issues/61509). Global properties are not handled by this workaround.